requirements.txt
aiogram
python-dotenv
requests
pytest
pytest-asyncio
pytest-cov
yt-dlp